refactor into one dynamic page instead of multiple pages
